+++ src/supremo_lite/__init__.py
@@ -36,7 +36,7 @@ from .personalize import (
     get_personal_genome,
     get_alt_sequences,
     get_ref_sequences,
-    get_pam_disrupting_personal_sequences,
+    get_pam_disrupting_alt_sequences,
     get_alt_ref_sequences
 )